Day 9 Task

+++ server/src/main/java/com/edutech/progressive/service/impl/CourseServiceImplJpa.java
@@ -7,10 +7,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.edutech.progressive.entity.Course;
+import com.edutech.progressive.exception.CourseAlreadyExistsException;
+import com.edutech.progressive.exception.CourseNotFoundException;
 import com.edutech.progressive.repository.CourseRepository;
 import com.edutech.progressive.service.CourseService;
 
 @Service
+
 public class CourseServiceImplJpa implements CourseService {
 
     private CourseRepository courseRepository;
@@ -35,13 +38,25 @@ public class CourseServiceImplJpa implements CourseService {
 
     public Course getCourseById(int courseId) throws Exception {
 
-        return courseRepository.findByCourseId(courseId);
+        return courseRepository.findById(courseId)
+
+                .orElseThrow(
+                        () -> new CourseNotFoundException("Course with ID " + courseId + " not found for deletion"));
 
     }
 
     @Override
 
     public Integer addCourse(Course course) throws Exception {
+
+        Course existingCourse = courseRepository.findByCourseName(course.getCourseName());
+
+        if (existingCourse != null) {
+
+            throw new CourseAlreadyExistsException(
+                    "Course with this name already exists, Course Name: " + course.getCourseName());
+
+        }
 
         return courseRepository.save(course).getCourseId();
 
@@ -51,6 +66,15 @@ public class CourseServiceImplJpa implements CourseService {
 
     public void updateCourse(Course course) throws Exception {
 
+        Course existingCourse = courseRepository.findByCourseName(course.getCourseName());
+
+        if (existingCourse != null && existingCourse.getCourseId() != course.getCourseId()) {
+
+            throw new CourseAlreadyExistsException(
+                    "Course with this name already exists, Course Name: " + course.getCourseName());
+
+        }
+
         courseRepository.save(course);
 
     }
@@ -59,9 +83,17 @@ public class CourseServiceImplJpa implements CourseService {
 
     public void deleteCourse(int courseId) throws Exception {
 
+        if (!courseRepository.existsById(courseId)) {
+
+            throw new CourseNotFoundException("Course with ID " + courseId + " not found for deletion");
+
+        }
+
         courseRepository.deleteById(courseId);
 
     }
+
+    @Override
 
     public List<Course> getAllCourseByTeacherId(int teacherId) {
